Added content back with HTML tags filtering
Hexo excludes HTML/XML tags, and their attributes, from data being indexed and searchable. So we can easily escape all of them. But it's better to format the content after that, so I added lots of typograf rules here.

Please note that you can't use this file on your site because of `filter(item => item.path.indexOf('help') !== -1)`. Remove this string if necessary.